Prevented generation of same HTML code multiple times
Changed function getParentLayout(). 

Added a 'distinct' to the select to avoid repeated code generation for the same layoutid. This
arises when the interfacerole.xml file refers to the same layout id for multiple roles
which is quite possible since more than one role may be associated with the same layout.
The repeated code generation was leading to multiple generation of the same ../dwr/interface/PojoName.js
<script> statements resulting in non-working of DOM functions such as getElementById() which further
resulted in some interfaces not working.